requirements.txt
asgiref==3.4.1
Django==3.2.6
pytz==2021.1
sqlparse==0.4.1
psycopg2==2.9.1
PyJWT==2.1.0
djangorestframework==3.12.4

